fix(homeserver): fix narrowing conversion in admin_auth_denied status
admin_auth_denied computed the 401/403 status as an unsigned int ternary
(missing ? 401U : 403U) and passed it to response(), whose status parameter is
std::uint16_t. Under -Wconversion -Werror the runtime ternary triggered
-Wimplicit-int-conversion and broke the build. Cast the ternary result to
std::uint16_t explicitly.

+++ src/homeserver/local_http_router.cpp
@@ -178,7 +178,7 @@ namespace
             return std::nullopt;
         }
         auto const missing = admin.denial == AdminAuthResult::Denial::missing_token;
-        auto const status = missing ? 401U : 403U;
+        auto const status = static_cast<std::uint16_t>(missing ? 401 : 403);
         auto const body =
             missing ? std::string{"admin authentication required"} : std::string{"admin privileges required"};
         return response(status, body, observability_headers(correlation, "text/plain; charset=utf-8"));
